Version 1.22.0 alexa-client-sdk
Changes in this update:

Feature enhancements, updates, and resolved issues from all releases are available on the [Amazon developer portal](https://developer.amazon.com/docs/alexa/avs-device-sdk/release-notes.html)

+++ Captions/Implementation/src/SystemClockDelay.cpp
@@ -22,8 +22,8 @@ namespace alexaClientSDK {
 namespace captions {
 
 void SystemClockDelay::delay(std::chrono::milliseconds milliseconds) {
-    auto duration =
-        std::chrono::milliseconds(std::max(std::chrono::milliseconds::zero().count(), milliseconds.count()));
+    auto duration = std::chrono::milliseconds(std::max(
+        static_cast<int64_t>(std::chrono::milliseconds::zero().count()), static_cast<int64_t>(milliseconds.count())));
     std::this_thread::sleep_for(duration);
 }
